Resizer: return byte[]

+++ src/Services/Core/Models/ResizingRequest.cs
@@ -2,7 +2,7 @@ namespace PixieFit.Core.Models;
 
 public class ResizingRequest 
 {
-    public decimal Height { get; set; }
-    public decimal Width { get; set; }
+    public double Height { get; set; }
+    public double Width { get; set; }
     public byte[] PhotoBytes { get; set; }
 }
